fix: polish and move folder
