+++ packages/core/src/types.ts
@@ -13,7 +13,7 @@ import type * as errorTracking from '@dd/error-tracking-plugin';
 import type { RumOptions } from '@dd/rum-plugin/types';
 import type * as rum from '@dd/rum-plugin';
 import type { RumPrivacyOptions } from '@dd/rum-privacy-plugin/types';
-import type * as rumPrivacy from '@dd/rum-privacy-plugin';
+import type * as privacy from '@dd/rum-privacy-plugin';
 import type { TelemetryOptions } from '@dd/telemetry-plugin/types';
 import type * as telemetry from '@dd/telemetry-plugin';
 // #imports-injection-marker
@@ -233,7 +233,7 @@ export interface Options extends BaseOptions {
     // #types-injection-marker
     [errorTracking.CONFIG_KEY]?: ErrorTrackingOptions;
     [rum.CONFIG_KEY]?: RumOptions;
-    [rumPrivacy.CONFIG_KEY]?: RumPrivacyOptions;
+    [privacy.CONFIG_KEY]?: RumPrivacyOptions;
     [telemetry.CONFIG_KEY]?: TelemetryOptions;
     // #types-injection-marker
     customPlugins?: GetCustomPlugins;
