--- packages/plugins/rum-privacy/src/string-imports.d.ts
@@ -1,8 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed under the MIT License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2019-Present Datadog, Inc.
-
-declare module '*.js-txt' {
-    const file: string;
-    export default file;
-}
