+++ packages/plugins/rum/privacy/src/built/privacy-helpers/index.ts
@@ -5,15 +5,16 @@
 /* eslint-env browser */
 /* global globalThis */
 (globalThis as any).$DD_ALLOW = new Set();
-(globalThis as any).$DD_ALLOW_OBSERVERS = new Set<() => void>();
+// @ts-ignore
+(globalThis as any).$DD_ALLOW_OBSERVERS;
 
 export function $(newValues: string[] | TemplateStringsArray) {
     const initialSize = (globalThis as any).$DD_ALLOW.size;
     newValues.forEach((value) => (globalThis as any).$DD_ALLOW.add(value));
     if ((globalThis as any).$DD_ALLOW.size !== initialSize) {
-        (globalThis as any).$DD_ALLOW_OBSERVERS.forEach((cb: () => void) => {
-            cb();
-        });
+        if ((globalThis as any).$DD_ALLOW_OBSERVERS) {
+            (globalThis as any).$DD_ALLOW_OBSERVERS.forEach((cb: () => void) => cb());
+        }
     }
 
     return newValues;
