+++ packages/plugins/rum/privacy/tsconfig.json
@@ -1,5 +1,5 @@
 {
-    "extends": "../../../tsconfig.json",
+    "extends": "../../../../tsconfig.json",
     "compilerOptions": {
         "baseUrl": "./",
         "rootDir": "./",
